Restore ability to run with `SCOREP_TOTAL_MEMORY=4G` again
First, never use `assert` to validate user input.
Second, all checks were already in place in `SCOREP_Memory_Initialize`.
Third, though because of the degradation to `uint32_t` arguments they didn't had
any effect anymore.

Partly reverts revision:12440.

Subversion: revision:14089

+++ src/measurement/SCOREP_Memory.c
@@ -7,7 +7,7 @@
  * Copyright (c) 2009-2012,
  * Gesellschaft fuer numerische Simulation mbH Braunschweig, Germany
  *
- * Copyright (c) 2009-2012, 2014,
+ * Copyright (c) 2009-2012, 2014, 2019,
  * Technische Universitaet Dresden, Germany
  *
  * Copyright (c) 2009-2012,
@@ -95,8 +95,8 @@ static bool out_of_memory;
 static SCOREP_Allocator_PageManager* definitions_page_manager;
 
 void
-SCOREP_Memory_Initialize( uint32_t totalMemory,
-                          uint32_t pageSize )
+SCOREP_Memory_Initialize( uint64_t totalMemory,
+                          uint64_t pageSize )
 {
     if ( is_initialized )
     {
